last update only applied to integer stats, now params as well

+++ Core/SoarKernel/include/soar_module.h
@@ -200,7 +200,7 @@ namespace soar_module
 	};
 	
 	// these are easy definitions for int and double parameters
-	typedef primitive_param<int> integer_param;
+	typedef primitive_param<long> integer_param;
 	typedef primitive_param<double> decimal_param;
